fix(dnd): hide overflow
